rename utils module name in ontology app
