+++ ontology/utils.py
